Refactoring de los Controladores.

+++ TodoApi/Controllers/EventController.cs
@@ -1,14 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using TodoApi.Models;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace TodoApi.Controllers
- 
+
 {
     [Route("api/[controller]")]
     public class EventController : Controller
@@ -22,6 +21,7 @@ namespace TodoApi.Controllers
 
             if (_context.EventItems.Count() == 0)
             {
+                //Codigo a ser eliminado en futuros refactorings
                 Event toAdd = new Event();
                 toAdd.nombre = "pelicula";
                 _context.EventItems.Add(toAdd);
@@ -38,12 +38,12 @@ namespace TodoApi.Controllers
         //Arroja la excepcion si no encuentra el Evento con el Id seleccionado...Revisar Metodo. 
         // GET api/values/5
         [HttpGet("{id}")]
-        public Event GetById(int id)
+        public IActionResult GetById(int id)
         {
             Event toR = _context.EventItems.SingleOrDefault(c => c.ID == id);
             if (toR == null)
-                throw new NullReferenceException();
-            return toR;
+                return NotFound();
+            return new ObjectResult(toR);
         }
 
         // POST api/values
@@ -52,9 +52,8 @@ namespace TodoApi.Controllers
         {
             //si el item es nulo lo procesa como badRequest
             if (item == null)
-            {
                 return BadRequest();
-            }
+            
             //agrega el item evento a la base de datos
             _context.EventItems.Add(item);
             _context.SaveChanges();
@@ -65,11 +64,15 @@ namespace TodoApi.Controllers
         //Actualizo los campos del Evento dado. 
         // PUT api/values/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody]Event item)
+        public IActionResult Update(int id, [FromBody]Event item)
         {
+            //Si no existe Evento con el id dado en la base de datos
             Event toModify = _context.EventItems.SingleOrDefault(c => c.ID == id);
             if (toModify == null)
-                throw new NullReferenceException();
+                return NotFound();
+            //Si nos pasan un evento con un parametro nulo
+            if (item == null)
+                return BadRequest();
             toModify.nombre = item.nombre;
             toModify.organizacionID = item.organizacionID;//No debe de ser necesario, linea a eliminar en proximos refactorings
             toModify.img = item.img;
@@ -78,19 +81,24 @@ namespace TodoApi.Controllers
             toModify.description = item.description;
 
             //guardo los cambios en la base de datos...
-
+            _context.EventItems.Update(toModify);
             _context.SaveChanges();
+            return new NoContentResult();
 
 
         }
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
             Event toRemove = _context.EventItems.SingleOrDefault(c => c.ID == id);
+            //Si el evento con el id dado no se encuentra en la base de datos.
+            if (toRemove == null)
+                return NotFound();
             _context.EventItems.Remove(toRemove);
             _context.SaveChanges();
+            return new NoContentResult();
            
         }
     }
